+++ gtsam/constrained/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Install headers
 file(GLOB constraint_headers "*.h")
-install(FILES ${constraint_headers} DESTINATION include/gtsam/constraint)
+install(FILES ${constraint_headers} DESTINATION include/gtsam/constrained)
 
 # Build tests
 add_subdirectory(tests)
