+++ gtsam/constrained/NonlinearEqualityConstraint.cpp
@@ -15,7 +15,7 @@
  * @author  Yetong Zhang, Frank Dellaert
  * @date    Aug 3, 2024 */
 
-#include <gtsam/constraint/NonlinearEqualityConstraint.h>
+#include <gtsam/constrained/NonlinearEqualityConstraint.h>
 #include <gtsam/inference/FactorGraph-inst.h>
 
 namespace gtsam {
