+++ gtsam/constrained/tests/testNonlinearInequalityConstraint.cpp
@@ -20,7 +20,7 @@
 #include <gtsam/base/Testable.h>
 #include <gtsam/base/TestableAssertions.h>
 #include <gtsam/base/numericalDerivative.h>
-#include <gtsam/constraint/NonlinearInequalityConstraint.h>
+#include <gtsam/constrained/NonlinearInequalityConstraint.h>
 #include <gtsam/nonlinear/factorTesting.h>
 
 #include "constrainedExample.h"
